+++ src/main/java/dev/jaxydog/content/CustomContent.java
@@ -8,6 +8,7 @@ import dev.jaxydog.content.item.CustomItems;
 import dev.jaxydog.content.power.CustomActions;
 import dev.jaxydog.content.power.CustomConditions;
 import dev.jaxydog.content.power.CustomPowers;
+import dev.jaxydog.content.sound.CustomSoundEvents;
 import dev.jaxydog.register.ContentRegistrar;
 import dev.jaxydog.register.IgnoreRegistration;
 
@@ -28,5 +29,6 @@ public final class CustomContent extends ContentRegistrar {
 	public static final CustomActions ACTIONS = new CustomActions();
 	public static final CustomConditions CONDITIONS = new CustomConditions();
 	public static final CustomPowers POWERS = new CustomPowers();
+	public static final CustomSoundEvents SOUND_EVENTS = new CustomSoundEvents();
 
 }