+++ src/main/java/dev/jaxydog/content/sound/CustomSoundEvents.java
@@ -1,0 +1,9 @@
+package dev.jaxydog.content.sound;
+
+import dev.jaxydog.register.ContentRegistrar;
+
+public final class CustomSoundEvents extends ContentRegistrar {
+
+	public static final CustomSoundEvent SPRAY_BOTTLE_USE = new CustomSoundEvent("spray_bottle_use");
+
+}